+++ src/routes.php
@@ -2,131 +2,88 @@
 
 declare(strict_types = 1);
 
-use Illuminate\Support\Arr;
+use Illuminate\Routing\Router;
+use Rebing\GraphQL\GraphQL;
 use Rebing\GraphQL\GraphQLController;
 
-$router = app('router');
-$schemaParameterPattern = '/\{\s*graphql\_schema\s*\?\s*\}/';
+$routeConfig = config('graphql.route');
 
-$router->group(array_merge([
-    'prefix' => config('graphql.prefix'),
-    'middleware' => config('graphql.middleware', []),
-], config('graphql.route_group_attributes', [])), function ($router) use ($schemaParameterPattern): void {
-    /** @var \Illuminate\Routing\Router $router */
+if ($routeConfig) {
+    /** @var Router $router */
+    $router = app('router');
 
-    // Routes and controllers
-    $routes = config('graphql.routes');
-    $controllers = config('graphql.controllers', GraphQLController::class . '@query');
-
-    $queryTypesMap = [
-        'query' => [],
-        'mutation' => [],
-    ];
-
-    /** @var array $queryTypesMap */
-    $queryTypesMap = array_combine(
-        array_keys($queryTypesMap),
-        array_map(function (string $type) use ($routes, $controllers, $queryTypesMap) {
-            $row = $queryTypesMap[$type];
-            $row['route'] = is_array($routes) ? Arr::get($routes, $type) : $routes;
-            $row['controller'] = is_array($controllers) ? Arr::get($controllers, $type) : $controllers;
-
-            return $row;
-        }, array_keys($queryTypesMap))
+    $routeGroupAttributes = array_merge(
+        [
+            'prefix' => $routeConfig['prefix'] ?? 'graphql',
+            'middleware' => $routeConfig['middleware'] ?? [],
+        ],
+        $routeConfig['group_attributes'] ?? []
     );
 
-    // Specific query type routes
-    $queryTypesMapWithRoutes = array_filter($queryTypesMap, function (array $row, string $type) use ($queryTypesMap) {
-        if ('mutation' === $type && $queryTypesMap['mutation']['route'] === $queryTypesMap['query']['route']) {
-            return;
-        }
+    $router->group(
+        $routeGroupAttributes,
+        function (Router $router) use ($routeConfig): void {
+            $schemas = GraphQL::getNormalizedSchemasConfiguration();
+            $defaultSchema = config('graphql.default_schema', 'default');
 
-        return null !== $row['route'];
-    }, ARRAY_FILTER_USE_BOTH);
+            foreach ($schemas as $schemaName => $schemaConfig) {
+                // A schemaConfig can in fact be a \GraphQL\Type\Schema object
+                // in which case no further information can be extracted from it
+                if (is_object($schemaConfig)) {
+                    $schemaConfig = [];
+                }
 
-    if ($queryTypesMapWithRoutes) {
-        $schemaConfig = Rebing\GraphQL\GraphQL::getNormalizedSchemasConfiguration();
+                $actions = array_filter([
+                    'uses' => $schemaConfig['controller'] ?? $routeConfig['controller'] ?? GraphQLController::class . '@query',
+                    'middleware' => $schemaConfig['middleware'] ?? $routeConfig['middleware'] ?? null,
+                ]);
 
-        $defaultConfig = Arr::get($schemaConfig, config('graphql.default_schema'), []);
-        $defaultMiddleware = $defaultConfig['middleware'] ?? [];
-        $defaultMethod = $defaultConfig['method'] ?? ['get', 'post'];
+                // Add route for each schema…
+                $router->addRoute(
+                    ['GET', 'POST'],
+                    $schemaName,
+                    $actions + ['as' => "graphql.$schemaName"]
+                );
 
-        foreach ($queryTypesMapWithRoutes as $type => $info) {
-            if (preg_match($schemaParameterPattern, $info['route'])) {
-                foreach ($defaultMethod as $method) {
-                    $routeName = "graphql.$type";
-
-                    if ('get' !== $method) {
-                        $routeName .= ".$method";
-                    }
-                    $router->{$method}(
-                        preg_replace($schemaParameterPattern, '', $info['route']),
-                        [
-                            'uses' => $info['controller'],
-                            'middleware' => $defaultMiddleware,
-                            'as' => $routeName,
-                        ]
+                // … and the default schema against the group itself
+                if ($schemaName === $defaultSchema) {
+                    $router->addRoute(
+                        ['GET', 'POST'],
+                        '',
+                        $actions + ['as' => 'graphql']
                     );
                 }
-
-                foreach ($schemaConfig as $name => $schema) {
-                    foreach (Arr::get($schema, 'method', ['get', 'post']) as $method) {
-                        $routeName = "graphql.$name";
-
-                        if ('get' !== $method) {
-                            $routeName .= ".$method";
-                        }
-                        $route = $router->{$method}(
-                            Rebing\GraphQL\GraphQL::routeNameTransformer($name, $schemaParameterPattern, $info['route']),
-                            [
-                                'uses' => $info['controller'],
-                                'middleware' => Arr::get($schema, 'middleware', []),
-                                'as' => $routeName,
-                            ]
-                        );
-
-                        $route->where($name, $name);
-                    }
-                }
-            } else {
-                $router->get($info['route'], [
-                    'uses' => $info['controller'],
-                    'as' => "graphql.$type",
-                ]);
-                $router->post($info['route'], [
-                    'uses' => $info['controller'],
-                    'as' => "graphql.$type.post",
-                ]);
             }
         }
-    }
-});
+    );
+}
 
 if (config('graphql.graphiql.display', true)) {
-    $router->group([
-        'prefix' => config('graphql.graphiql.prefix', 'graphiql'),
-        'middleware' => config('graphql.graphiql.middleware', []),
-    ], function ($router) use ($schemaParameterPattern): void {
-        /** @var \Illuminate\Routing\Router $router */
-        $graphiqlController = config('graphql.graphiql.controller', GraphQLController::class . '@graphiql');
+    /** @var Router $router */
+    $router = app('router');
+    $graphiqlConfig = config('graphql.graphiql');
 
-        $graphiqlAction = ['uses' => $graphiqlController];
+    $router->group(
+        [
+            'prefix' => $graphiqlConfig['prefix'] ?? 'graphiql',
+            'middleware' => $graphiqlConfig['middleware'] ?? [],
+        ],
+        function (Router $router) use ($graphiqlConfig): void {
+            $actions = [
+                'uses' => $graphiqlConfig['controller'] ?? GraphQLController::class . '@graphiql',
+            ];
 
-        foreach (config('graphql.schemas') as $name => $schema) {
-            $route = $router->get(
-                Rebing\GraphQL\GraphQL::routeNameTransformer($name, $schemaParameterPattern, '{graphql_schema?}'),
-                $graphiqlAction + ['as' => "graphql.graphiql.$name"]
-            );
-            $route->where($name, $name);
+            // A graphiql route for each schema…
+            /** @var string $schemaName */
+            foreach (array_keys(config('graphql.schemas')) as $schemaName) {
+                $router->get(
+                    $schemaName,
+                    $actions + ['as' => "graphql.graphiql.$schemaName"]
+                );
+            }
 
-            $route = $router->post(
-                Rebing\GraphQL\GraphQL::routeNameTransformer($name, $schemaParameterPattern, '{graphql_schema?}'),
-                $graphiqlAction + ['as' => "graphql.graphiql.$name.post"]
-            );
-            $route->where($name, $name);
+            // … and one for the default schema against the group itself
+            $router->get('/', $actions + ['as' => 'graphql.graphiql']);
         }
-
-        $router->get('/', $graphiqlAction + ['as' => 'graphql.graphiql']);
-        $router->post('/', $graphiqlAction + ['as' => 'graphql.graphiql.post']);
-    });
+    );
 }